backend/requirements.txt
annotated-types==0.7.0
blinker==1.9.0
click==8.2.0
Flask==3.0.0
Flask-Cors==4.0.0
itsdangerous==2.2.0
Jinja2==3.1.6
MarkupSafe==3.0.2
pydantic==2.7.0
pydantic_core==2.18.1
PyYAML==6.0
typing_extensions==4.13.2
Werkzeug==3.1.3
